Change column lengths to match json schema
The password field is ok, as it is stored hashed.
The hash uses bcrypt, which produces 60 character long hash.

One more consideration is the actual plain text password length. As
bcrypt only uses the first 72 bytes, the length might be good to limit
to more closely reflect the actual usable length.