Update customer stakes visualization
Customers' stakes are represented by Dots, which display a tooltip of
the customer stakes data on hover.

A single customer Dot diameter is calculated from the max diameter and
the percentage of value received from the milestone. Minimun diameter
multiplier is 0.2 so dots are clearly visible.

Max diameter scales according to the amount of customers related to
the milestones. Currently the max diameter range is [60px, 120px]
(120px is achieved with 7 customers).