Modify client weighting
Added a description to client weights page.
Adjusted the range of the slider to [0,2]. When set to zero, the
the client's ratings are ignored.
(effects are not implemented yet)
Modified database to use float type for customer weight.
Reset and save buttons are now disabled when no changes to customer
weight have been made.