Rewrite sorting
The Sort type defines the comparison to use, and optionally a sorting
key.

The key may be a function, in which case the sorting calls that function
once for each item to obtain key to use for comparison. The key may also be
non-optional property of the sorted object, in which case that property is used for
comparison.