Move task map edges back on the z-axis
Previously the edges were in front of the task groups. Thus the edges
could obstruct a task making it unreadable. The edge's hit box is also
larger than the visual size, and can block interactions on a task below
it.

Now the edges are rendered below the task groups. The task group
background has some transparency to make it easier to see, where the
edge connects to. The existing hover effect also helps tracing an
individual edge, even if it goes under task groups.