Add task attachments API improvements
Main improvements are:

Fetch attachments when fetching tasks to avoid requesting attachments
every time task overview is opened. This elimates the need for a single
GET tasks -route.

Attachments are posted in the same request when adding a task, so
everything can be done in a single query.

Add emitRoadmapEvents to task attachments -routes, so other users fetch
the new data too.